Re-arrange workers, and add checks for common logger errors
