+++ apps/greencheck/models.py
@@ -1,3 +1,4 @@
+from dataclasses import dataclass
 import decimal
 import ipaddress
 import logging
@@ -34,6 +35,25 @@ logger = logging.getLogger(__name__)
 # wait for reply on these.
 - greenenergy - also an old table
 """
+
+
+@dataclass
+class SiteCheck:
+    """
+    A representation of the Sitecheck object from the PHP app.
+    We use it as a basis for logging to the Greencheck, but also maintaining
+    our green_domains tables.
+    """
+
+    url: str
+    ip: str
+    data: bool
+    green: bool
+    hosting_provider_id: int
+    checked_at: str
+    match_type: str
+    match_ip_range: int
+    cached: bool
 
 
 class IpAddressField(Field):
